requirement.txt
cffi==1.17.1
cryptography==44.0.0
mpmath==1.3.0
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==24.3.0
sympy==1.13.3
